separate include paths for client/server/shared, fix includes in all files
