+++ include/server/gameobject.hpp
@@ -1,9 +1,9 @@
 #pragma once
 
-#include "shared/base_gameobject.hpp"
-#include "shared/core.hpp"
-#include "shared/transform.hpp"
-#include "shared/types.hpp"
+#include "base_gameobject.hpp"
+#include "core.hpp"
+#include "transform.hpp"
+#include "types.hpp"
 
 #include <memory>
 #include <string>
