+++ src/server/gameobject.cpp
@@ -1,4 +1,4 @@
-#include "server/gameobject.hpp"
+#include "gameobject.hpp"
 
 void GameObject::applyMovement(const glm::vec3 &direction) {
   glm::vec3 pos = transform->getPosition();
