+++ src/shared/CMakeLists.txt
@@ -5,7 +5,7 @@ collect_sources("${CMAKE_CURRENT_SOURCE_DIR}" SHARED_SRCS)
 add_library(${LIB_NAME} STATIC ${SHARED_SRCS})
 
 target_include_directories(${LIB_NAME} PRIVATE
-    ${INCLUDE_DIRECTORY}
+    ${SHARED_INCLUDE_DIRECTORY}
 )
 
 target_link_libraries(${LIB_NAME} PRIVATE
