+++ src/shared/objects/animal.cpp
@@ -1,22 +1,22 @@
-#include "shared/objects/animal.hpp"
+// #include "objects/animal.hpp"
 
-Animal::Animal(float x, float y, float z, const string &type)
-    : positionX(x), positionY(y), positionZ(z), animalType(type) {}
+// Animal::Animal(float x, float y, float z, const string &type)
+//     : positionX(x), positionY(y), positionZ(z), animalType(type) {}
 
-Animal::~Animal() {}
+// Animal::~Animal() {}
 
-void Animal::move(float deltaX, float deltaY, float deltaZ) {
-  positionX += deltaX;
-  positionY += deltaY;
-  positionZ += deltaZ;
-}
+// void Animal::move(float deltaX, float deltaY, float deltaZ) {
+//   positionX += deltaX;
+//   positionY += deltaY;
+//   positionZ += deltaZ;
+// }
 
-void Animal::interact(const GameObject &other) {}
+// void Animal::interact(const GameObject &other) {}
 
-float Animal::getPositionX() const { return positionX; }
+// float Animal::getPositionX() const { return positionX; }
 
-float Animal::getPositionY() const { return positionY; }
+// float Animal::getPositionY() const { return positionY; }
 
-float Animal::getPositionZ() const { return positionZ; }
+// float Animal::getPositionZ() const { return positionZ; }
 
-string Animal::getAnimalType() const { return animalType; }
+// string Animal::getAnimalType() const { return animalType; }
